Move derived state into dedicated folder
