+++ src/clientState/derived/selectedSatelliteVariableObject.ts
@@ -5,10 +5,10 @@
 
 import {selector} from 'recoil';
 
-import selectedSatelliteVariableAtom from './selectedSatelliteVariable';
-import {ISatelliteVariableOptions, ISatelliteVariableIndex} from '../types/query/satelliteVariables';
-import {queryClient} from '../util/query';
-import {SERVERSTATE_KEY_VARIABLES_INDEX} from '../serverState/variablesIndex';
+import selectedSatelliteVariableAtom from '../selectedSatelliteVariable';
+import {ISatelliteVariableOptions, ISatelliteVariableIndex} from '../../types/query/satelliteVariables';
+import {queryClient} from '../../util/query';
+import {SERVERSTATE_KEY_VARIABLES_INDEX} from '../../serverState/variablesIndex';
 
 
 // TODO: Should we call selectors "atoms" for simplicity? They are used
